+++ packages/webgpu/src/hooks.ts
@@ -4,6 +4,14 @@ import type { RNCanvasContext, CanvasRef, NativeCanvas } from "./Canvas";
 
 type Unsubscribe = () => void;
 
+export const warnIfNotHardwareAccelerated = (adapter: GPUAdapter) => {
+  if (adapter.info.architecture === "swiftshader") {
+    console.warn(
+      "GPUAdapter is not hardware accelerated. This is common on Android emulators. Rendering will be slow.",
+    );
+  }
+};
+
 export const useSurface = () => {
   const [surface, setSurface] = useState<NativeCanvas | null>(null);
   const ref = useCanvasEffect(() => {
@@ -11,6 +19,36 @@ export const useSurface = () => {
     setSurface(sur);
   });
   return { ref, surface };
+};
+
+export const useDevice = (
+  adapterOptions?: GPURequestAdapterOptions,
+  deviceDescriptor?: GPUDeviceDescriptor,
+) => {
+  const [adapter, setAdapter] = useState<GPUAdapter | null>(null);
+  const [device, setDevice] = useState<GPUDevice | null>(null);
+  useEffect(() => {
+    (async () => {
+      if (!adapter) {
+        const adp = await navigator.gpu.requestAdapter(adapterOptions);
+        if (!adp) {
+          throw new Error("No appropriate GPUAdapter found.");
+        }
+        setAdapter(adp);
+        warnIfNotHardwareAccelerated(adp);
+        return;
+      }
+      if (!device) {
+        const dev = await adapter.requestDevice(deviceDescriptor);
+        if (!dev) {
+          throw new Error("No appropriate GPUDevice found.");
+        }
+        setDevice(dev);
+        return;
+      }
+    })();
+  }, [adapter, adapterOptions, device, deviceDescriptor]);
+  return { adapter, device };
 };
 
 export const useGPUContext = () => {
